Add support for copy/pasting inline images in Xlib
Images can be pasted from the clipboard if the owner can provide them in
an image/png format.  No support for incremental pasting yet so large
images do not work.  There seems to be a problem with RGB/BGR encoding
but it only affects the preview, and shows up fine in other clients
(tested with venom).