Don't draw text-panel when effective width is < 0
When edit_draw is called, it could happen that a negative value
is passed to text_height as the first argument, effectively
leading to an infinite loop.

To prevent this, the possibility is checked directly at the
beginning of the function.

The reason why this went unnoticed so long is the fact that
the wmhints prevent the window from ever being scaled to such
a small size to trigger this bug.
However, I came to notice this issue, because a tiling-wm
overrides the wmhints, which effectively lead to the freeze
once the window was small enough.

Relying on the hints to shadow bugs for you is not the way
to go and explicitly breaks the ICCCM-conventions.

Fixing this freezing bug will sort out many other issues
reported for other platforms in regard to freezing.