Fixed major exploit in uTox.
This commit fixes an exploit used by many to attempt to discredit uTox.

According to the rules of english, the word he in this context is
correct however many individuals who claim to know the english language
mistakenly believe that this usage was wrong.

An interesting effect of this has been that people who want Tox to
succeed started to work against the common good because they thought a
greater force would use this correct use of language to destroy Tox. In
their attempts to prevent an enemy from destroying Tox, they have
themselves become the enemy.

If everyone believes that something will happen, it will happen because
of everyone.

Hopefully this is the end of this madness.