Make icon compatible with WinXP
The icon failed to render under WinXP both
as an application icon in explorer and as a tray icon.
Just opened and saved it again in GIMP and that fixed it.
Don't know what exactly was wrong, the only difference
I noticed is 256 -> 16 color depth change. But since
the image consists just of the black color of varying
transparency, I think it's fine.