Rewrite context menus with i18n support
CONTEXTMENU struct is generalized by introducing
ondisplay() callback for fetching menu item strings,
and adding a userdata field.

One out-of-the-box CONTEXT variety is provided.

It treats userdata as an array of UI_STRING_IDs
to be displayed (beware, contextmenu_new() function
doesn't copy the array, so make sure it has proper lifespan).

All previously existed context menus are replaced with it.