Fixed multiple misuses of snprintf.
If the generated string is longer than the output buffer, snprintf
will not overflow the buffer - this is known. It also returns the
number of bytes that would have been written. This may be
counterintuitive: it may seem like it should return the number of
bytes actually written. It does not.