Refactored utox default post message thread...
It's almost pointless at this point, a lot of the messages were straight up removed, and
even more were replaced with a simple redraw() call.

At some point we should decide if we want to just strip this out in favor of tox_postmessage(), or]
build in some other use for this function.