Dissabled logging by default
New installs will no longer log conversations.

Logging seems innapropiate for privacy-oriented software. It is my opinion that
logging is not a feature with much benifit to most users but it puts them all
at risk. It also seems rude to log the words of whoever you are talking to
without first asking their permission.