Change translate for 'Tox ID is self ID'
ID means '账号' in Chinese is well-known. Did 'Tox ID' need translated to'Tox 账号‘, I'm not sure. As a proper noun, I didn't translate it.
But, 'Tox ID is self ID', More popular translation is 'Tox ID 是你自己的账号'.
'Tox ID 是你自己的 ID', looks like 'Tai Shan is a high Shan', it's not very authentic rather than 'Tai Shan is a high mountain' or 'Tai Shan mountain is a high mountain'.
But it does not impede understanding.