Replace all instances of SWITCH_ICON with SWITCH_INNER
According to the mockups, switches have icons. This change should
make distinguishing between actual icons and the inner parts a bit
easier.

On a side note, I'm not sure if INNER is the best name for that
part of the switch, but I couldn't come up with anything better.
Suggestions welcome.