MAJOR REFACTOR
full on refactor of uTox, with the header guards written by @endoffile78 I was finally
able to remove the insane dependency every file had for . Now nearly every
file should only be including the bare minimum needed to compile and run.

There are a few outstanding issues with the additional partial conversions of i18n_*
from uint8_t to char, but none of them should be too bad to finally fixup.

Finally, I added a smattering of extra warnings so there's the constant reminder that
we need to stop using static struct arrays for friends and groups. As part of that
fix, we'll likely need to work through and remove the usage of flist_get_selected()
but none of that should be too hard.