Refactor postmessages()
This change makes it more clear what needs to be done inside the
UI thread, vs what's allowed to be done by the toxcore thread.

In this case, redraw() must only be called from the UI thread.
While this is always the way it should have been, the addition
of the 2,3... windows each having their own drawing device
context removes the flexablity we could allow before.