Fixed Hanging GUI when changing network settings
In case of an error in `init_toxcore()` currently `tox_thread_init`
is not set to `1` causing the thread calling `tox_settingschanged()` to wait forever in
[this loop](https://github.com/uTox/uTox/blob/1835eae7a96a1335f781308b0afa730ab3647483/src/tox.c#L211-L213).

There are two error situations, the first is simple, waiting for
password input before tox can connect. The second is fatal error when
utox is unable to create a tox instance. The last one was failing here.
It is now fixed by handling this case in an exctra condition in
`toxcore_thread()`.

I have also fixed the error condition in `init_toxcore()` to fail only
if `*tox == NULL`, the other parts of the condition make no sense.

Additionally the Toggle of the UDP button got inverted when setting a
proxy, this is fixed by updating the button at the same time when
updating the setting.

Also disabling the UDP button when enabling proxy makes sense to avoid
creating impossible config combination.

Finally in the error sitatuation a link is created from the "Not
Connected" message to the network tab to allow adjusting the settings
quickly.