proper include of freetype in cmake
learned something new today. CMake has a powerful tool for finding
libraries on different platforms. Instead of hardcoding an include dir,
we can let cmake determine the correct paths and also let it fail with a
proper error (i.e. "freetype librariy is required") instead of cryptic
linker/compiler errors if freetype is not available.

Could be done for more libraries but freetype was the one that needed
special handling so I added it for freetype only for now.