Refactor the init system in uTox
This introduces a regression in --theme=name, it'll be overwritten what
exists in the savefile. I'm not sure what the best fix is. But I do know
that parse_args() must be called before utox_init().

The solution likely is to refactor all the settings.* changes to a
seperate function

Signed-off-by: Gregory Mullen (grayhatter) <greg@grayhatter.com>