tray icon fix in d505cb8f was not correct, cleanup
Commit d505cb8f tried to read the icon_size directly from
_binary_icons_utox_128x128_png_size which is wrong.
The way it was done before via &_binary_icons_utox_128x128_png_size is
the correct way to do it.

Unfortunately this is not working anymore on systems with PIE (position
independent executable) enabled by default. This causes the loader to
relocate the absolute symbol _binary_icons_utox_128x128_png_size, which
defeats the purpose of the symbol being absolute.
(See discussion here: [1])

A proper fix would be not to use ld to generate the object file but to
serialize the png file to C code via a helper programm during the build
process. I will look into that.

For now just calculate the icon size from the symbols
_binary_icons_utox_128x128_png_start and
_binary_icons_utox_128x128_png_end.

[1] https://www.reddit.com/r/C_Programming/comments/71qx7t/reading_the_size_of_linked_binary_object_file/)