Fix minor issue with HOME and END keys
This fixes a non-standard behaviour of the HOME and END keys.
Now, when HOME/END is pressed while there is also a selection, the cursor goes to the beginning or end of line, instead of only to the beginning or end of the selection. Going to the beginning/end of the selection is correct behaviour for LEFT and RIGHT, but not for HOME/END.

This also fixes a premature break bug.