Create a compilation database by default
This is useful for Clang-based tools, in particular language servers like
clangd. I think it's a good idea to provide it by default to improve the
development experience.  I didn't find a good way to detect a development
setup, so I just made it the default and disable it in scripts that are
mostly for CI. Sadly this turned out a bit crazy :(

If CMake read environment variables like Make, then a developer could use
"export CMAKE_EXPORT_COMPILE_COMMANDS=ON" (once, for all projects).
Is there a proper way to do something like that in CMake?