Fix double free in flist when changing settings w/ empty flist
This pretty much only happens if you double-click e.g. the udp toggle to
trigger a restart of tox twice in rapid succession. The first restart
calls flist_freeall() which empties the flist and leaves old pointers in
`item` and `shown_list` and the second flist_freeall() will try to free
the old pointers again.

[avoidr: return early]